Basic UI

+++ app/src/main/java/com/example/lab5/PageFormat.kt
@@ -2,5 +2,5 @@ package com.example.lab5
 
 class PageFormat (
     val Name: String,
-    val Price: Float
+    val Price: Int,
 )
